requirements.txt
streamlit>=1.30.0
streamlit-authenticator>=0.2.3
groq>=0.4.0
plotly>=5.18.0
pandas>=2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
pyyaml>=6.0.0
